Organize base/styles like the other features
Now that the other features have a way to switch between Native and Web
and use an import-export logic based on features rather than on files,
make base/styles consistent with the other features.
