+++ features/base/styles/components/native/index.js
@@ -1,0 +1,1 @@
+export * from './styles/ColorPalette';
