+++ features/base/styles/components/native/styles/ColorPalette.js
@@ -1,11 +1,11 @@
 /**
  * The application color palette.
  */
-var ColorPalette = {
+const ColorPalette = {
     buttonUnderlay: '#495258',
     jitsiBlue: '#00ccff',
     jitsiRed: '#ff0000',
     jitsiToggled: this.buttonUnderlay
 }
 
-export default ColorPalette;
+export ColorPalette;
